SQLite Refactor, User Caching, FFC Cleanup
FFC really sucked. I don’t even know why I cached stuff in `.json` files before.
User profiles are now cached in SQLite for better maintainability, fewer bugs, and fewer crashes. I will slowly phase out FFC at this point, since it has become one of the most pointless things I’ve made.
